HACK: inject activation key into cachi2 stage
Credit to
https://github.com/brianwcook/build-definitions/tree/STONEBLD-2618, but
that branch had a lot of stuff on it and was using old images that
couldn't handle gomod very well. If it turns out we can't use activation
keys, we could do the same for the certificate. I'd add them in the
options, but there's not a way to mount them up right now without
mangling the pipeline task.